using default route for login, register, and errors
